Refactor demo file structure (#35)
